Making query tuple count accurate
Previously we were subtracting 1 from the output of a query that ran as
below:

`tuples=$(psql -A -q -t -P pager=off -v ON_ERROR_STOP=OFF -v EXPLAIN_ANALYZE="" -f test2.sql | wc -l)`

With this change we are running the queries like this:

`tuples=$(psql -A -q -t -P pager=off -v ON_ERROR_STOP=ON -v EXPLAIN_ANALYZE="" -f test2.sql | wc -l; exit ${PIPESTATUS[0]})`

This allows us to check for error exit on a query in case of failures or
timeouts. Also we do not need to do any math on tuples.

Even if a query fails for whatever reason, we still continue to the next
query in line.

Signed-off-by: Ekta Khanna <ekhanna@pivotal.io>